requirements.txt
dash>=3.2.0
dash-bootstrap-components>=2.0.4
plotly>=6.3.0
pandas>=2.3.2
yfinance>=0.2.65
requests>=2.32.5
beautifulsoup4>=4.13.5
perplexityai>=0.12.0
puthon-dotenv>=1.1.1
httpx>=0.28.1